Fixed small issue with Wyatt's item pickup system
The gameobject used for the held item's position relative to the character was a child of the GFX, so when the player was holding an item and crouched, the item would "crouch" too and lose half its height. It could then be dropped in that form to permanently make it smaller. I just moved that gameobject up in the hierarchy and the issue was resolved.